Hide "Public" group for logged-out users under certain conditions
As discussed in https://github.com/hypothesis/product-backlog/issues/702,
hide the "Public" group in the client if:

 - The user is logged out
 - The page has associated groups
 - The user has not visited a direct link to an annotation in the
   "Public" group

This is implemented by making the `store.allGroups()` selector used by
the group list exclude the "Public" group under the above conditions.

A `directLinkedAnnotationId` field has been added to the selection state
to support this. This field is initialized from the settings passed
through to the sidebar app when it starts.