Remember the direct linked annotation's group after it is unloaded
This fixes the following issue:

 1. As a logged-out user, follow a link to a direct-linked annotation in
    "Public" on a page that has associated groups.
 2. Switch from "Public" group to another group.

    Expected: "Public" group remains in group list.
    Actual: "Public" group disappears from group list.

The "Public" group was disappearing because the group ID of the
direct-linked annotation was no longer known after the DL'd annotation
was unloaded.

This is fixed by recording the group ID when the DL'd annotation is
loaded and not resetting it when the DL'd annotation is unloaded.